refactor: Extract SwipeAction enum to its own file
